backend/pyproject.toml
[tool.poetry]
name = "fast-template"
version = "0.1.0"
description = "Modular FastApi Template"
authors = ["peplxx"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
python-multipart = "^0.0.17"
psycopg2-binary = "^2.9.10"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
